fix(testcontainers): author shard sequencer as CJS so config loads on node 20
