+++ internal-packages/testcontainers/package.json
@@ -7,7 +7,10 @@
   "exports": {
     ".": "./src/index.ts",
     "./webapp": "./src/webapp.ts",
-    "./sequencer": "./src/sequencer.ts"
+    "./sequencer": {
+      "types": "./src/sequencer.d.cts",
+      "default": "./src/sequencer.cjs"
+    }
   },
   "dependencies": {
     "@clickhouse/client": "^1.11.1",
